FSA-5467: Removed all price checks and separate tests for integrations
