+++ projects/fsastorefront-e2e-cypress/cypress/helpers/changeRequest.ts
@@ -22,20 +22,6 @@ export function checkChangeMileageSteps() {
     });
 }
 
-export function enterNewMileage() {
-  cy.get('cx-fs-cms-form-submit')
-    .should('be.visible')
-    .within(() => {
-      cy.get('h3').contains('Auto Information');
-      cy.get('[name="vehicleAnnualMileage"]').type(50000);
-    });
-}
-
-export function checkChangedPolicyPremium() {
-  cy.get('.offset-1').contains(currentDate);
-  cy.get('.col-3.semi-bold').eq(2).contains(' €100.00 / Monthly ');
-}
-
 export function checkChangeRequestConfirmation() {
   cy.get('cx-fs-change-process-confirmation').should('be.visible');
   cy.get('h2').contains(' Change Request Confirmation ');
@@ -54,30 +40,6 @@ export function startChangeCoverage() {
   cy.get('.link.position-absolute').contains('Edit').click();
 }
 
-export function checkOptionalExtras() {
-  cy.get('cx-fs-change-coverage')
-    .should('be.visible')
-    .within(() => {
-      cy.get('.section-header-heading').contains(' Optional Extras ');
-      cy.get('.row.mx-3').should('have.length', 4);
-      cy.get('h6').eq(0).contains('Uninsured Coverage');
-      cy.get('h6').eq(1).contains('Roadside Assistance');
-      cy.get('h6').eq(2).contains('Trailer Liability');
-      cy.get('h6').eq(3).contains('Winter Tires');
-    });
-}
-
-export function addRoadsideAssistance() {
-  cy.get('.secondary-button').should('have.length', 4).contains(' Add ');
-  cy.get('.row.mx-3')
-    .eq(1)
-    .within(() => {
-      cy.get('h6').contains('Roadside Assistance');
-      cy.get('.secondary-button').click();
-      cy.get('.secondary-button').contains('Remove');
-    });
-}
-
 export function checkChangeCoverageSteps() {
   cy.get('h2').contains(' Change Auto Insurance Policy ');
   cy.get('.action-button').should('contain', 'Cancel');
@@ -89,26 +51,33 @@ export function checkChangeCoverageSteps() {
     });
 }
 
-export function checkChangedPolicyNewPremium() {
-  cy.get('.offset-1').contains(currentDate);
-  cy.get('.col-4.semi-bold').eq(5).contains(' €100.00 / Monthly ');
-  cy.get('.col-3.semi-bold').eq(2).contains(' €100.00 / Monthly ');
+export function enterNewMileage() {
+  cy.get('cx-fs-cms-form-submit')
+    .should('be.visible')
+    .within(() => {
+      cy.get('[name="vehicleAnnualMileage"]').type('50000');
+    });
 }
 
-export function checkInboxMessages() {
-  cy.get('.fs-tab').should('be.visible').contains('Auto');
-  cy.get('.fs-tab')
+export function checkOptionalExtrasBronze() {
+  cy.get('cx-fs-change-coverage')
     .should('be.visible')
-    .contains('Policies')
-    .click({ force: true });
-  cy.get('.message')
-    .should('be.visible')
-    .click({ multiple: true })
     .within(() => {
-      cy.contains('Policy ID');
-      cy.contains('Dear Ben Moore, The change request');
-      cy.contains(todaysDate);
-      cy.get('.box-shadow').should('be.visible');
+      cy.get('.section-header-heading').contains(' Optional Extras ');
+      cy.get('.row.mx-3').should('have.length', 3);
+      cy.get('h6').eq(0).contains('Collision Coverage');
+      cy.get('h6').eq(1).contains('Uninsured Coverage');
+      cy.get('h6').eq(2).contains('Trailer Liability');
     });
-  cy.get('.message').should('have.class', 'read');
+}
+
+export function addTrailerLiability() {
+  cy.get('.secondary-button').should('have.length', 3);
+  cy.get('.row.mx-3')
+    .eq(2)
+    .within(() => {
+      cy.get('h6').contains('Trailer Liability');
+      cy.get('.secondary-button').click();
+      cy.get('.secondary-button').contains('Remove');
+    });
 }
