+++ projects/fsastorefront-e2e-cypress/cypress/integration/checkout/autoe2e/autoBronze_Invoice_ChangeCoverage.e2e.ts
@@ -6,17 +6,12 @@ import * as checkout from '../../../helpers/checkout/checkoutSteps';
 import { selectPaymentMethodInvoice } from '../../../helpers/checkout/insurance/payment';
 import * as myPolicies from '../../../helpers/my-account/policies';
 import * as changeRequest from '../../../helpers/changeRequest';
-import {
-  addTrailerLiability,
-  checkChangedCoveragePremium,
-  checkOptionalExtrasBronze,
-} from '../../../helpers/checkout/insurance/policyChange_integration';
 
 Cypress.config('defaultCommandTimeout', 500000);
 
 context('Auto Bronze Checkout with change coverage', () => {
   before(() => {
-    cy.visit('http://10.27.241.80/financial/en/EUR/');
+    cy.visit('/');
   });
 
   it('Should register a new user', () => {
@@ -35,8 +30,7 @@ context('Auto Bronze Checkout with change coverage', () => {
   });
 
   it('Should check comparison table and select main and optional products', () => {
-    autoIntegration.checkAutoComparisonTableAudi();
-    autoIntegration.selectAutoBronzeAudi();
+    auto.selectAutoBronze();
     auto.checkOptionalProductsBronze();
     checkout.clickContinueButton();
   });
@@ -47,14 +41,12 @@ context('Auto Bronze Checkout with change coverage', () => {
     auto.populateVehicleDetails();
     auto.populateMainDriverData();
     auto.populateAdditionalData();
-    autoIntegration.checkAutoBrozneAudiMiniCart();
     checkout.clickContinueButton();
   });
 
   it('Should bound a quote', () => {
     checkout.checkCheckoutStep('Your Auto Insurance', '7');
     checkout.checkProgressBarInsurance();
-    autoIntegration.checkAutoBrozneAudiMiniCart();
     checkout.checkAccordions('generalQuoteAccordions');
     checkout.clickContinueButton();
     checkout.ConfirmBindQuote();
@@ -70,25 +62,25 @@ context('Auto Bronze Checkout with change coverage', () => {
 
   it('Should check my policies and policy details page', () => {
     //waiting for replication process to be completed
-    cy.wait(35000);
+    cy.wait(200000);
     myPolicies.checkMyPoliciesPage();
-    autoIntegration.checkReplicatedBronzeA5Policy();
+    autoIntegration.checkReplicatedPolicy('Bronze');
     cy.get('.overview-section-title').contains(' Auto Insurance Policy ');
-    checkout.checkAccordions('integrationPolicyDetails');
+    //This should be included for integration tests
+    //checkout.checkAccordions('integrationPolicyDetails');
   });
 
   it('Should complete change coverage checkout', () => {
     changeRequest.startChangeCoverage();
     //check change coverage - first step
     changeRequest.checkChangeCoverageSteps();
-    checkOptionalExtrasBronze();
+    changeRequest.checkOptionalExtrasBronze();
     //check continue button is disabled if coverage is not added
     cy.get('.primary-button').contains('Continue').should('be.disabled');
-    addTrailerLiability();
+    changeRequest.addTrailerLiability();
     checkout.clickContinueButton();
     //check change preview - second step
     changeRequest.checkChangeCoverageSteps();
-    checkChangedCoveragePremium();
     cy.get('.primary-button').should('contain', 'Submit').click();
     changeRequest.checkChangeRequestConfirmation();
   });
